pyproject.toml
[project]
name = "robotics-2037"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "squigglepy",
]

[tool.uv.sources]
squigglepy = { git = "https://github.com/danwahl/squigglepy.git", branch = "update-python-3.13" }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
pythonpath = ["."]
